complex to float

+++ experiments/earthtopography/plot.py
@@ -53,7 +53,7 @@ truth = pyssht.inverse(pys2let.lm_hp2lm(map2alm(topo, L - 1), L), L, Reality=Tru
 
 MAP_idx = np.where(logpi == max(logpi))
 MAP = file["predictions"][MAP_idx][0]
-MAP = MAP.reshape((L, 2 * L - 1))
+MAP = MAP.reshape((L, 2 * L - 1)).astype(float)
 MAP_plt, _ = pyssht.mollweide_projection(MAP, L)
 maxapost = plotting.plot_map(MAP_plt, title="Maximum a posetriori solution")
 maxapost.savefig(filename("MAP"))
